Abstracted level concept to level template and modified game to manage levels - vchadha
