+++ src/com/base/game/Game.java
@@ -4,148 +4,64 @@ import com.base.engine.*;
 import com.base.game.gameobject.entity.Player;
 import com.base.game.gameobject.entity.Boss;
 import com.base.game.interfaces.UI;
-import com.base.game.utilities.Delay;
-import com.base.game.utilities.LevelTransition;
-import org.lwjgl.opengl.GL11;
+import com.base.game.levels.Level;
+import com.base.game.levels.LevelTransition;
 
 import java.awt.*;
-import java.io.IOException;
 import java.util.ArrayList;
-
-import static org.lwjgl.opengl.GL11.*;
+import java.util.List;
 
 public class Game {
     public static Game game;
-    private ArrayList<GameObject> gameObjects;
-    private ArrayList<GameObject> toAdd;
-    private ArrayList<GameObject> toRemove;
-
-    private Player player;
-    private Boss boss;
-    private UI ui;
-
-    private boolean levelOver;
-    private boolean gameOver;
-    private LevelTransition lvlTransition;
+    private ArrayList<Level> levels;
+    private int currLevel;
 
     public Game() {
-        gameObjects = new ArrayList<>();
-        toAdd = new ArrayList<>();
-        toRemove = new ArrayList<>();
+        currLevel = 0;
+        levels = new ArrayList();
 
-        player = new Player(Display.getWidth() / 2 - 30, Display.getHeight() / 2 - 30, 41, 82, "./res/player.png", 4f, 20, 5);
+        Level level1 = new Level("./res/bricks.jpg");
+        levels.add(level1);
+    }
 
-        boss = new Boss(Display.getWidth() / 2 - 35, Display.getHeight() - 150, 70, 70, "", 2f,67, 5);
-
-        addObj(player);
-        addObj(boss);
-
-        ui = new UI(player.getHealth(), boss.getHealth());
-
-        levelOver = false;
-        gameOver = false;
-        lvlTransition = new LevelTransition();
+    private Level getCurrLevel() {
+        return levels.get(currLevel);
     }
 
     public void update() {
-        if(!levelOver) {
-            for (GameObject object : gameObjects) {
-                if (!object.isRemoved())
-                    object.update();
-                else
-                    toRemove.add(object);
-            }
-            if (ui != null) {
-                ui.update();
-            }
-
-            if (!toAdd.isEmpty()) {
-                while (!toAdd.isEmpty())
-                    gameObjects.add(toAdd.remove(0));
-            }
-
-            if (!toRemove.isEmpty()) {
-                for (GameObject object : toRemove) {
-                    gameObjects.remove(object);
-                }
-
-                toRemove.clear();
-            }
-        }
+        levels.get(currLevel).update();
     }
 
     public void render() {
-        gameObjects.forEach(GameObject::render);
-        if (ui != null) { ui.render(); }
-        if(levelOver)
-        {
-            if(lvlTransition != null && lvlTransition.render(gameOver))
-            {
-                lvlTransition = null;
-                levelOver = false;
-            }
-        }
+        levels.get(currLevel).render();
     }
 
     public void addObj(GameObject obj) {
-        toAdd.add(obj);
+        getCurrLevel().addObj(obj);
+    }
+
+    public void endLevel() {
+        getCurrLevel().endLevel();
+    }
+
+    public ArrayList<GameObject> getCloseObjects(GameObject object, float range) {
+        return getCurrLevel().getCloseObjects(object, range);
+    }
+
+    public void levelOver(boolean lose) {
+        getCurrLevel().levelOver(lose);
     }
 
     public float getPlayerX() {
-        return player.getX();
+        return getCurrLevel().getPlayerX();
     }
 
     public float getPlayerY() {
-        return player.getY();
+        return getCurrLevel().getPlayerY();
     }
 
     public int getHealth(boolean isPlayer)
     {
-        return (isPlayer ? player.getHealth() : boss.getHealth());
-    }
-
-    public ArrayList<GameObject> getCloseObjects(GameObject object, float range)
-    {
-        ArrayList<GameObject> closeObj = new ArrayList<>();
-        float p1 = object.getX() - (object.getWidth()/2f) - range;
-        float p2 = object.getY() - (object.getHeight()/2f) - range;
-        Rectangle field = new Rectangle((int)p1, (int)p2, (int)(object.getWidth() + 2 * range), (int)(object.getHeight() + 2 * range));
-        /*
-        glBegin(GL11.GL_QUADS);
-            glVertex2f(field.x, field.y);
-            glVertex2f(field.x + field.width, field.y);
-            glVertex2f(field.x + field.width, field.y + field.height);
-            glVertex2f(field.x, field.y + field.height);
-        glEnd();
-        */
-        for(GameObject o : gameObjects)
-        {
-            if(!o.equals(object) && Physics.checkCollision(field, o))
-                closeObj.add(o);
-        }
-        return closeObj;
-    }
-
-    public void levelOver(boolean lose)
-    {
-        levelOver = true;
-        gameOver = lose;
-        lvlTransition.init();
-    }
-
-    public void endLevel()
-    {
-        boolean done = false;
-        while(!done)
-        {
-            try {
-                gameObjects.remove(1); // Get rid of everything besides the player
-            }
-            catch(IndexOutOfBoundsException e) {
-                done = true;
-            }
-        }
-        ui = null;
-        // TODO Create door
+        return getCurrLevel().getHealth(isPlayer);
     }
 }
